Declare the nginx-ingress app
The roundtable kustomization's role is to declare the basic resources
and Argo CD applications of the Roundtable cluster.

To start, we make roundtable declare the nginx-ingress Application for
Argo CD. In turn, that application, references the off-the-shelf helm
chart customization in deployments/nginx-ingress.

We create namespaces separately so that the Application resource already
has the namespace.

This is patterned after
https://github.com/argoproj/argoproj-deployments/blob/master/argoproj/kustomization.yaml