Add IgnoreExtraneous annotations to PVCs
Strimzo 0.14.0 has added the ability to pass annotations to to
PersistentVolumeClaims of Kafka and Zookeeper. See
https://strimzi.io/docs/latest/#con-customizing-template-properties-str
We can use this to apply an IgnoreExtraneous compare option to the PVCs
so that ArgoCD doesn't attempt to sync them (they're controlled by the
Kafka operator, not by any original chart).

This should also improve the cluster stability and prevent persistent
volumes from being recreated and rebound as the PVCs get bounced around
by Argo competing with the Strimzi operator.

For DM-21529