[DM-23266] Move Vault dependencies to security app
To allow the security app-of-apps to be self-contained as the first
step of a new cluster bootstrap, move the dependencies of Vault into
that app-of-apps.

Merge the namespace configurations into the application configurations
for the applications that are moved.

Add the cert-manager-crd application to the same file as the
cert-manager application, since this is an internal implementation
detail of cert-manager.