[DM-23286] Work around nginx clusterIP bug
The nginx-ingress Helm chart has reintroduced an issue with
clusterIP on upgrades that was previously fixed.  The clusterIP
has to be hard-coded to successfully upgrade unless the prior
version was installed with omitClusterIP, which it wasn't for our
configuration.

Hard-code the existing clusterIPs for now until upstream finds a
better fix.