[DM-23547] Upgrade to cert-manager 0.13.1
Import new CRDs from the upstream Git repository and update the
cert-manager-crd app to deploy them.  Add some operational docs for
why this is a separate app and how it is updated.

Update all issuers to use the new API group and the new syntax for
specifying the solver.

Update all annotations to use the new annotation label for specifying
the issuer.

Remove the Argo CD ignoreDifferences configuration to see if this
problem is fixed with the latest cert-manager.

Remove the disable-validation annotation for the cert-manager
namespace since that upstream documentation page no longer exists
and upstream no longer appears to recommend this change.