Add GitHub workflow for Helm chart linting
Lint all the embedded Helm charts with cart-testing-action, although
be generous in not requiring them to follow the schema because they're
normally stub charts.

Add Dependabot configuration to keep the actions up to date.

Remove an old copy of the Vault Helm chart, left over from a previous
iteration of deploying Vault.