Deploy Gafaelfawr on Roundtable
This will take over the /login, /logout, and /auth routes, which
should be merged with the routes declared by the Roundtable ingress
and inherit the TLS settings from them.

Point at the current development branch since this will be tested as
an OpenID Connect provider for Open Distro for Elasticsearch.