Restore proxy Kibana authentication, document OpenID
Switch back to proxy authentication for Kibana since it's easier
to use with roles, but add tested documentation for how to use
OpenID Connect should it ever be useful in the future.

Revert to the default Gafaelfawr deployment, since we don't need
the new endpoint for proxy authentication.  We'll pick it up later
with the next regular Gafaelfawr release.