Replace prometheus with kube-prometheus-system
The older Prometheus chart has been deprecated and the new recommended
Helm chart is kube-prometheus-system.  Replace the old Prometheus
configuration with the new chart.  Since the data from the current
system is not currently that significant, this uninstalls the old
Prometheus install and creates a new one rather than trying to preserve
and move the persistent volumes.

Add a VaultSecret for the Grafana GitHub OAuth 2 secret, which was
previously using a manually-created secret.  (The relevant secret has
already been stored in Vault.)