Run documentation builds on PRs
Match what we're doing with Phalanx.  The specific problem this
fixes is that updating the GitHub Actions configuration for the
docs build in a PR branch didn't retry the docs build.

Also add the GitHub Actions configuration to the triggering paths,
which would have fixed the same problem a different way.